Move `Track` to separate file
